Fiz muitas alteraçoes, Implementei o uso de DTOs no metodo put produto, para fazer alteraçoes nos produtos

+++ CadastroProdutoMySQL/Modelos/Estoque.cs
@@ -4,14 +4,14 @@ namespace CadastroProdutoMySQL.Modelos
 {
     public class Estoque
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         public int Quantidade { get; set; }
 
         public Estoque()
         {
         }
 
-        public Estoque(int id, int quantidade)
+        public Estoque(long id, int quantidade)
         {
             Id = id;
             Quantidade = quantidade;
